feat: package weread2notion action
