+++ src/weread2notion/__init__.py
@@ -1,0 +1,3 @@
+"""WeRead to Notion sync package."""
+
+__version__ = "1.0.0"
